[feat] Finish lexer implementation.
The lexer is now more or less finished. I'm not sure if I should bother trying to improve the performance. It takes about 90ms to tokenize the prelude.

Additionaly a REPL and the option to run file has been added. `cargo run` for the REPL and `cargo run -- [FILE]` for selecting a file as input.